requirements.txt
# Gmail MCP Server Dependencies

# MCP Framework
mcp>=1.0.0

# Data validation
pydantic>=2.0.0

# Google API libraries
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.100.0
